refactor(diagnostic): reuse ByteLocator for per-source line lookup
